repositories in data access layer

+++ Poizon.DAL/Interfaces/IAvailabilityRepository.cs
@@ -4,6 +4,6 @@ namespace Poizon.DAL.Interfaces;
 
 public interface IAvailabilityRepository : IBaseRepository<Availability>
 {
-    Task<int> GetCountByClothesId(int id);
+    Task<int> GetCountByClothesId(long id);
 }
