+++ Poizon.DAL/Interfaces/IClothesRepository.cs
@@ -4,13 +4,12 @@ namespace Poizon.DAL.Interfaces;
 
 public interface IClothesRepository : IBaseRepository<Clothes>
 {
-    Task<IEnumerable<Clothes>> GetClothesByCategoryId(int id);
-    Task<IEnumerable<Clothes>> GetClothesBySubCategoryId(int id);
-    Task<IEnumerable<Clothes>> GetClothesByBrandId(int id);
-    Task<IEnumerable<Clothes>> GetClothesByModelId(int id);
-    Task<IEnumerable<Clothes>> GetClothesByStyleId(int id);
-    Task<IEnumerable<Clothes>> GetClothesBySexId(int id);
-    Task<IEnumerable<Clothes>> GetClothesBySizeId(int id);
-    Task<IEnumerable<Clothes>> GetClothesByCost(int minCost, int maxCost);
+    Task<IEnumerable<Clothes>> GetClothesByCategoryId(long id);
+    Task<IEnumerable<Clothes>> GetClothesBySubCategoryId(long id);
+    Task<IEnumerable<Clothes>> GetClothesByBrandId(long id);
+    Task<IEnumerable<Clothes>> GetClothesByModelId(long id);
+    Task<IEnumerable<Clothes>> GetClothesByStyleId(long id);
+    Task<IEnumerable<Clothes>> GetClothesBySexId(long id);
+    Task<IEnumerable<Clothes>> GetClothesBySizeId(long id);
+    Task<IEnumerable<Clothes>> GetClothesByCost(long minCost, long maxCost);
 }
-
